Added solution - LeetHub

+++ Chocolate Distribution Problem - GFG/chocolate-distribution-problem.cpp
@@ -7,8 +7,8 @@ class Solution{
     public:
     long long findMinDiff(vector<long long> a, long long n, long long m){
         sort(a.begin(),a.end());
-        long long diff=a[m-1]-a[0];
-        for(long long i=m;i<n;i++){
+        int diff=a[m-1]-a[0];
+        for(int i=m;i<n;i++){
             if(diff>(a[i]-a[i-m+1])) diff=a[i]-a[i-m+1];
         }
         return diff;
